Wiring client

+++ backend/migrations/20260409160950_privacypass_tokens_operations.up.sql
@@ -10,7 +10,7 @@ CREATE TABLE as_token_allowance(
     user_uuid uuid PRIMARY KEY,
     user_domain TEXT NOT NULL,
     operation_type SMALLINT NOT NULL,
-    remaining INTEGER NOT NULL,
+    remaining SMALLINT NOT NULL,
     epoch SMALLINT NOT NULL,
     FOREIGN KEY (user_uuid, user_domain) REFERENCES as_user_record (user_uuid, user_domain) ON DELETE CASCADE,
     CONSTRAINT unique_user_operation UNIQUE (user_uuid, user_domain, operation_type)
